--- src/main/java/Quampus/demo/login/config/data/RedisCustomService.java
@@ -1,11 +1,0 @@
-package Quampus.demo.login.config.data;
-
-import java.util.Set;
-
-public interface RedisCustomService {
-    void saveRedisData(String keyName, String value, Long limitTime) ;
-    String getRedisData(String key);
-    void deleteRedisData(String key);
-    boolean hasKey(String key);
-    Set<String> getKeysByPattern(String pattern);
-}
