--- src/main/java/Quampus/demo/login/controller/UserController.java
@@ -1,4 +1,0 @@
-package Quampus.demo.login.controller;
-
-public class UserController {
-}
