--- src/main/java/Quampus/demo/login/dto/ReasonDTO.java
@@ -1,14 +1,0 @@
-package Quampus.demo.login.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@Builder
-public class ReasonDTO {
-    private HttpStatus httpStatus;
-    private final boolean isSuccess;
-    private final String code;
-    private final String message;
-}
