--- src/main/java/Quampus/demo/login/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package Quampus.demo.login.repository;
-
-import Quampus.demo.login.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByEmail(String email);
-}
