--- src/main/java/Quampus/demo/login/service/UserService.java
@@ -1,7 +1,0 @@
-package Quampus.demo.login.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserService {
-}
