+++ src/main/java/swyp/qampus/login/controller/UserController.java
@@ -1,0 +1,4 @@
+package swyp.qampus.login.controller;
+
+public class UserController {
+}
