+++ src/main/java/swyp/qampus/login/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package swyp.qampus.login.repository;
+
+import swyp.qampus.login.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByEmail(String email);
+}
