+++ src/main/java/swyp/qampus/login/service/OauthService.java
@@ -1,0 +1,12 @@
+package swyp.qampus.login.service;
+
+
+import swyp.qampus.login.entity.User;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import jakarta.servlet.http.HttpServletResponse;
+
+public interface OauthService {
+    User oAuthLogin(String code, HttpServletResponse httpServletResponse);
+
+    void kakaoLogout(String accessToken) throws JsonProcessingException;
+}
