+++ src/main/java/swyp/qampus/login/service/UserService.java
@@ -1,0 +1,7 @@
+package swyp.qampus.login.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserService {
+}
